backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Utils
python-dotenv==1.0.0

# Payments (YooKassa)
yookassa==3.1.0

# Image processing
Pillow==10.2.0

# Push notifications
pywebpush==1.14.0

# Email
aiosmtplib==3.0.1




